Move HMAC to new files
