+++ lib/crypto/c_src/Makefile.in
@@ -76,6 +76,7 @@ CRYPTO_OBJS = $(OBJDIR)/crypto$(TYPEMARKER).o \
 	$(OBJDIR)/bn$(TYPEMARKER).o \
 	$(OBJDIR)/digest$(TYPEMARKER).o \
 	$(OBJDIR)/engine$(TYPEMARKER).o \
+	$(OBJDIR)/hmac$(TYPEMARKER).o \
 	$(OBJDIR)/rsa$(TYPEMARKER).o
 CALLBACK_OBJS = $(OBJDIR)/crypto_callback$(TYPEMARKER).o
 NIF_MAKEFILE = $(PRIVDIR)/Makefile
